microservices/shared/requirements.txt
Django==3.2.23
djangorestframework==3.14.0
django-cors-headers==4.3.1
PyJWT==2.8.0
requests==2.31.0
psycopg2-binary==2.9.9
celery==5.3.4
redis==5.0.1
django-extensions==3.2.3
Pillow==10.1.0
pytz==2024.1
sqlparse==0.4.4
asgiref==3.7.2 